warning fix --kazu

+++ UserDev/SelectionTool/OpT0Finder/Algorithms/LightPath.cxx
@@ -74,7 +74,7 @@ namespace flashana {
     QCluster_t result;
     result.clear();
 
-    for (int i = 0; i < trj.size() - 1; i++) {
+    for (size_t i = 0; i < trj.size() - 1; i++) {
       auto const& this_loc(trj[i]);
       auto const& last_loc(trj[i + 1]);
